requirements.txt
psutil
python-dotenv
pillow